[fix] 강제 언래핑, 파일 분리 등의 리뷰 반영
