+++ Project17-C-Map/InteractiveClusteringMap/Map/MapInteractor.swift
@@ -11,7 +11,6 @@ protocol ClusterBusinessLogic: class {
     
     func fetch(boundingBoxes: [CLong: BoundingBox], zoomLevel: Double)
     func remove(tileIds: [CLong])
-//    func add(coordinate: Coordinate)
     func delete(coordinate: Coordinate)
     
 }
@@ -57,10 +56,6 @@ final class MapInteractor: ClusterBusinessLogic {
         }
     }
     
-//    func add(coordinate: Coordinate) {
-//        presenter.add(poi: POI(x: coordinate.x, y: coordinate.y, id: coordinate.id, name: "", imageUrl: "", category: ""))
-//    }
-//
     func delete(coordinate: Coordinate) {
         quadTreeClusteringService.delete(coordinate: coordinate)
         presenter.delete(coordinate: coordinate)
