+++ Project17-C-Map/InteractiveClusteringMap/Map/MapViewController+NMFMapViewCameraDelegate.swift
@@ -27,7 +27,6 @@ extension MapViewController: NMFMapViewCameraDelegate {
     func mapView(_ mapView: NMFMapView, cameraDidChangeByReason reason: Int, animated: Bool) {
         self.transparentLayer?.sublayers?.forEach { subLayer in
             self.setMarkerPosition(marker: subLayer)
-            subLayer.removeAllAnimations()
         }
     }
